+++ POS/POS.Data/Config/CashLimitConfig.cs
@@ -1,0 +1,9 @@
+﻿using POS.Core.Config;
+
+namespace POS.Infrastructure.Config
+{
+    public class CashLimitConfig : ICashLimit
+    {
+        public decimal AddCashLimit { get ; set ; }
+    }
+}