+++ POS/POS/Modules/Payout/PayoutModule.cs
@@ -19,6 +19,10 @@ namespace POS.Modules.Payout
 
             var voucherSettings = configuration.GetConfigurationSection<VoucherSettingsFileConfig>("UserVoucherSettings");
             services.AddSingleton<IVoucherSettings>(voucherSettings);
+            var cashLimit = configuration.GetConfigurationSection<CashLimitConfig>("CashLimit");
+            services.AddSingleton<ICashLimit>(cashLimit);
+            var approvalRequired = configuration.GetConfigurationSection<NeedApprovalConfig>("NeedsApproval");
+            services.AddSingleton<INeedApproval>(approvalRequired);
         }
     }
 }