+++ POS/POS/Modules/Payout/ViewModels/TransactionViewModel.cs
@@ -20,6 +20,8 @@ using System;
 using System.Collections.ObjectModel;
 using System.Threading.Tasks;
 using System.Windows.Input;
+using POS.Core.Config;
+using System.Windows;
 
 namespace POS.Modules.Payout.ViewModels
 {
@@ -45,7 +47,8 @@ namespace POS.Modules.Payout.ViewModels
             IUserSession userSession,
             ILogEventDataService logEventDataService,
             IPrintService printService,
-            ILastReceiptService lastReceiptService)
+            ILastReceiptService lastReceiptService,
+            INeedApproval settings)
         {
             _messageBoxService = messageBoxService;
             _session = session;
@@ -58,10 +61,19 @@ namespace POS.Modules.Payout.ViewModels
             _serviceLocator = serviceLocator;
             _systemContext = systemContext;
             _eventAggregator = eventAggregator;
-
+            NeedsApproval = settings.IsApprovalRequired;
             _transaction = new Transaction();
         }
 
+        private bool _needApproval;
+        public bool NeedsApproval
+        {
+            get => _needApproval; set
+            {
+                _needApproval = value;
+                NotifyOfPropertyChange(nameof(NeedsApproval));
+            }
+        }
         public void AddTransactionItem(VoucherDto voucherDto)
         {
             try